pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "stock-movers"
version = "0.1.0"
description = "Daily stock movers intelligence tool with AI-powered catalyst summaries"
requires-python = ">=3.10"
dependencies = [
    "yfinance>=0.2.40",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
movers = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
